Add author-reported shift results to range shifts

The BioShifts package standardizes author-reported data in several forms to rates of change in the latitudinal or elevational direction over time. This function supplements the calculated shift rates with shift findings as reported by original study authors.

Usage

add_author_reported(data)

Arguments

data

input shifts dataset from get_shifts() function.

Value

Range shifts dataset with added columns containing author-reported shift values, and the figure, table, or dataset in the original publication from which the shift values were pulled.

Examples

get_shifts() |> add_author_reported()
#> # A tibble: 31,759 × 19
#>    id         article_id poly_id method_id eco   type  param sp_name_publication
#>    <chr>      <chr>      <chr>   <chr>     <chr> <chr> <chr> <chr>              
#>  1 A001_P1_E… A001       P1      M01       Ter   ELE   O     Aegithalos_caudatus
#>  2 A001_P1_E… A001       P1      M01       Ter   ELE   O     Certhia_familiaris 
#>  3 A001_P1_E… A001       P1      M01       Ter   ELE   O     Dendrocopos_major  
#>  4 A001_P1_E… A001       P1      M01       Ter   ELE   O     Dryocopus_martius  
#>  5 A001_P1_E… A001       P1      M01       Ter   ELE   O     Erithacus_rubecula 
#>  6 A001_P1_E… A001       P1      M01       Ter   ELE   O     Fringilla_coelebs  
#>  7 A001_P1_E… A001       P1      M01       Ter   ELE   O     Garrulus_glandarius
#>  8 A001_P1_E… A001       P1      M01       Ter   ELE   O     Nucifraga_caryocat…
#>  9 A001_P1_E… A001       P1      M01       Ter   ELE   O     Parus_ater         
#> 10 A001_P1_E… A001       P1      M01       Ter   ELE   O     Parus_caeruleus    
#> # ℹ 31,749 more rows
#> # ℹ 11 more variables: sp_name_checked <chr>, subsp_or_pop <chr>,
#> #   calc_rate <dbl>, calc_unit <chr>, direction <chr>, author_reported <chr>,
#> #   author_reported_unit <chr>, author_reported_sig <chr>,
#> #   author_reported_magnitude <dbl>, author_reported_angle <dbl>,
#> #   author_source <chr>