Add methods to range shift documentations

add_methods supplements the minimal shifts database with methodological information describing the methods by which individal shifts were documented. These include variables such as sampling years, definitions of range parameters, and more.

Usage

add_methods(data)

Arguments

data

input data from get_shifts().

Value

original shifts dataset, supplemented with methodological parameters for each individual range shift detection. See readme for details on methods columns.

Examples

get_shifts() |> add_methods() |> dplyr::glimpse()
#> Rows: 31,761
#> Columns: 29
#> $ id                           <chr> "A001_P1_ELE_O_M01", "A001_P1_ELE_O_M01",…
#> $ article_id                   <chr> "A001", "A001", "A001", "A001", "A001", "…
#> $ poly_id                      <chr> "P1", "P1", "P1", "P1", "P1", "P1", "P1",…
#> $ method_id                    <chr> "M01", "M01", "M01", "M01", "M01", "M01",…
#> $ eco                          <chr> "Ter", "Ter", "Ter", "Ter", "Ter", "Ter",…
#> $ type                         <chr> "ELE", "ELE", "ELE", "ELE", "ELE", "ELE",…
#> $ param                        <chr> "O", "O", "O", "O", "O", "O", "O", "O", "…
#> $ sp_name_publication          <chr> "Aegithalos_caudatus", "Certhia_familiari…
#> $ sp_name_checked              <chr> "Aegithalos_caudatus", "Certhia_familiari…
#> $ subsp                        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ calc_rate                    <dbl> -2.2128, -0.5106, -7.8723, -3.2340, 4.851…
#> $ calc_unit                    <chr> "m/year", "m/year", "m/year", "m/year", "…
#> $ direction                    <chr> "Lower Elevation", "Lower Elevation", "Lo…
#> $ sublocality_or_pop           <chr> "Giffre Valley", "Giffre Valley", "Giffre…
#> $ start_firstperiod            <dbl> 1978, 1978, 1978, 1978, 1978, 1978, 1978,…
#> $ midpoint_firstperiod         <dbl> 1979.0, 1979.0, 1979.0, 1979.0, 1979.0, 1…
#> $ end_firstperiod              <dbl> 1980, 1980, 1980, 1980, 1980, 1980, 1980,…
#> $ start_secondperiod           <dbl> 2001, 2001, 2001, 2001, 2001, 2001, 2001,…
#> $ midpoint_secondperiod        <dbl> 2001.5, 2001.5, 2001.5, 2001.5, 2001.5, 2…
#> $ end_secondperiod             <dbl> 2002, 2002, 2002, 2002, 2002, 2002, 2002,…
#> $ n_periods                    <dbl> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…
#> $ duration                     <dbl> 23.5, 23.5, 23.5, 23.5, 23.5, 23.5, 23.5,…
#> $ grain_size                   <chr> "small", "small", "small", "small", "smal…
#> $ sampling                     <chr> "TWO", "TWO", "TWO", "TWO", "TWO", "TWO",…
#> $ category                     <chr> "Survey-Resurvey", "Survey-Resurvey", "Su…
#> $ obs_type                     <chr> "occurrence", "occurrence", "occurrence",…
#> $ uncertainty_distribution     <chr> "Resurveyed", "Resurveyed", "Resurveyed",…
#> $ position_definition          <chr> "Number of occurrences and mean elevation…
#> $ position_definition_category <chr> "Mean", "Mean", "Mean", "Mean", "Mean", "…