Add taxonomy to range shifts

add_taxo() merges taxonomic classification to range shifts uploaded with the get_shifts() function.

Usage

add_taxo(data)

Arguments

df

input shifts dataset from get_shifts() function.

Value

Range shifts dataset with added columns containing taxonomic classification information for every species name, as written in the original publication.

Examples

get_shifts() |> add_taxo() |> dplyr::glimpse()
#> Rows: 31,759
#> Columns: 19
#> $ id                  <chr> "A001_P1_ELE_O_M1", "A001_P1_ELE_O_M1", "A001_P1_E…
#> $ article_id          <chr> "A001", "A001", "A001", "A001", "A001", "A001", "A…
#> $ poly_id             <chr> "P1", "P1", "P1", "P1", "P1", "P1", "P1", "P1", "P…
#> $ method_id           <chr> "M01", "M01", "M01", "M01", "M01", "M01", "M01", "…
#> $ eco                 <chr> "Ter", "Ter", "Ter", "Ter", "Ter", "Ter", "Ter", "…
#> $ type                <chr> "ELE", "ELE", "ELE", "ELE", "ELE", "ELE", "ELE", "…
#> $ param               <chr> "O", "O", "O", "O", "O", "O", "O", "O", "O", "O", …
#> $ sp_name_publication <chr> "Aegithalos_caudatus", "Certhia_familiaris", "Dend…
#> $ sp_name_checked     <chr> "Aegithalos_caudatus", "Certhia_familiaris", "Dend…
#> $ subsp_or_pop        <chr> "Giffre Valley", "Giffre Valley", "Giffre Valley",…
#> $ calc_rate           <dbl> -2.2128, -0.5106, -7.8723, -3.2340, 4.8511, -1.319…
#> $ calc_unit           <chr> "m/year", "m/year", "m/year", "m/year", "m/year", …
#> $ direction           <chr> "Lower Elevation", "Lower Elevation", "Lower Eleva…
#> $ kingdom             <chr> "Animalia", "Animalia", "Animalia", "Animalia", "A…
#> $ phylum              <chr> "Chordata", "Chordata", "Chordata", "Chordata", "C…
#> $ class               <chr> "Aves", "Aves", "Aves", "Aves", "Aves", "Aves", "A…
#> $ family              <chr> "Aegithalidae", "Certhiidae", "Picidae", "Picidae"…
#> $ taxon_id            <chr> "GBIF:2495000", "GBIF:2487603", "GBIF:2477968", "G…
#> $ taxon_method        <chr> "Find_Sci_Names", "Find_Sci_Names", "Find_Sci_Name…