Add author-reported shift results to range shifts

The BioShifts package standardizes author-reported data in several forms to rates of change in the latitudinal or elevational direction over time. This function supplements the calculated shift rates with shift findings as reported by original study authors.

Usage

add_author_reported(data)

Arguments

data

input shifts dataset from get_shifts() function.

Value

Range shifts dataset with added columns containing author-reported shift values, and the figure, table, or dataset in the original publication from which the shift values were pulled.

Examples

get_shifts() |> add_author_reported() |> dplyr::glimpse()
#> Rows: 31,759
#> Columns: 19
#> $ id                        <chr> "A001_P1_ELE_O_M1", "A001_P1_ELE_O_M1", "A00…
#> $ article_id                <chr> "A001", "A001", "A001", "A001", "A001", "A00…
#> $ poly_id                   <chr> "P1", "P1", "P1", "P1", "P1", "P1", "P1", "P…
#> $ method_id                 <chr> "M01", "M01", "M01", "M01", "M01", "M01", "M…
#> $ eco                       <chr> "Ter", "Ter", "Ter", "Ter", "Ter", "Ter", "T…
#> $ type                      <chr> "ELE", "ELE", "ELE", "ELE", "ELE", "ELE", "E…
#> $ param                     <chr> "O", "O", "O", "O", "O", "O", "O", "O", "O",…
#> $ sp_name_publication       <chr> "Aegithalos_caudatus", "Certhia_familiaris",…
#> $ sp_name_checked           <chr> "Aegithalos_caudatus", "Certhia_familiaris",…
#> $ subsp_or_pop              <chr> "Giffre Valley", "Giffre Valley", "Giffre Va…
#> $ calc_rate                 <dbl> -2.2128, -0.5106, -7.8723, -3.2340, 4.8511, …
#> $ calc_unit                 <chr> "m/year", "m/year", "m/year", "m/year", "m/y…
#> $ direction                 <chr> "Lower Elevation", "Lower Elevation", "Lower…
#> $ author_reported           <chr> "-52", "-12", "-185", "-76", "114", "-31", "…
#> $ author_reported_unit      <chr> "m", "m", "m", "m", "m", "m", "m", "m", "m",…
#> $ author_reported_sig       <chr> "N", "N", "Y", "N", "Y", "N", "N", "N", "Y",…
#> $ author_reported_magnitude <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ author_reported_angle     <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ author_source             <chr> "Table 3 pg 143", "Table 3 pg 143", "Table 3…