Add baseline temperatures of species- or study-area polygons for each range shift

Add baseline temperatures of species- or study-area polygons for each range shift

Usage

add_baselines(
  data,
  type = "SA",
  stat = c("mean"),
  exp = c("temp"),
  res = c(LAT = "25km", ELE = "1km")
)

Arguments

data

input data from get_shifts()

type

Choice of baseline temperatures from study area (SA) or species area (SP) polygons.

stat

Statistic of the given variable. Choices are min, 1Q, median, mean, 3Q, max.

exp

Exposure variable; Choices are "temp" (temperature") or "precip" (precipitation), or both.

res

Calculation resolution. Baseline temperatures in each species/study area were calculated with environmental raster layers at up to four resolutions: 1km, 25km, 50km, 110km, resulting in slightly different values. Choose a specific res ("1km","25km","50km","110km"), or use "best" to ensure each shift has a matching temperature (see vignette)

Value

Shifts database supplemented with selected temperature and/or precipitation baseline values within the study area or species-specific study area.

Examples

get_shifts() |> add_baselines()
#> Error in map(.x, .f, ...): ℹ In index: 1.
#> Caused by error in `dplyr::mutate(dplyr::left_join(data_split[[.x]], dplyr::select(baselines,
#>     article_id, poly_id, type, method_id, all_of(cols[[.x]])), by = dplyr::join_by(
#>     article_id, poly_id, method_id, type)), baseline_res = res[[.x]]) %>% dplyr::rename_at(
#>     dplyr::all_of(cols[[.x]]), function(col) stringr::str_replace(col, "_res.*",
#>       ""))`:
#> ! could not find function "%>%"
get_shifts(eco = "Mar") |> add_baselines(res = "25km")
#> Error in get_shifts(eco = "Mar"): unused argument (eco = "Mar")